Rename cmd package to command
